Add 2 kinds of events manager and 2 kinds of plugins managers
The WorldPluginsManager manages the plugins of one particular world. The ServerPluginsManager manages the ServerPlugins, which are loaded on the entire server or in multiple worlds. Note that a WorldPluginsManager also reports the ServerPlugins that are loaded in its world.

The WorldEventsManager manages events in one particular world. The ServerEventsManager manages events for the ServerPlugins. Actually, the ServerEventsManager is just a front-end to all the WorldEventsManagers, a convenient way to register/unregister events in every world where a ServerPlugin is loaded.